Fixed error when lock file exists

+++ head/spine/core.py
@@ -616,4 +616,5 @@ class Spine:
             if self.use_lock:
                 os.remove(lockfn)
                 logger.info('Removed lock at %s.' % lockfn)
-        self.wsServer.terminate()
+        if hasattr(self,'wsServer'):
+            self.wsServer.terminate()
